+++ cmd/zar/import/command.go
@@ -5,18 +5,12 @@ import (
 	"flag"
 	"fmt"
 	"os"
-	"path/filepath"
 
 	"github.com/alecthomas/units"
 	"github.com/brimsec/zq/archive"
 	"github.com/brimsec/zq/cmd/zar/root"
-	"github.com/brimsec/zq/pkg/bufwriter"
-	"github.com/brimsec/zq/pkg/fs"
-	"github.com/brimsec/zq/pkg/nano"
-	"github.com/brimsec/zq/zbuf"
 	"github.com/brimsec/zq/zio"
 	"github.com/brimsec/zq/zio/detector"
-	"github.com/brimsec/zq/zio/zngio"
 	"github.com/brimsec/zq/zng/resolver"
 	"github.com/mccanne/charm"
 )
@@ -44,7 +38,6 @@ type Command struct {
 	*root.Command
 	root        string
 	thresh      string
-	quiet       bool
 	empty       bool
 	ReaderFlags zio.ReaderFlags
 }
@@ -53,14 +46,9 @@ func New(parent charm.Command, f *flag.FlagSet) (charm.Command, error) {
 	c := &Command{Command: parent.(*root.Command)}
 	f.StringVar(&c.root, "R", os.Getenv("ZAR_ROOT"), "root directory of zar archive for chopped files")
 	f.StringVar(&c.thresh, "s", units.Base2Bytes(archive.DefaultConfig.LogSizeThreshold).String(), "target size of chopped files, as '10MB' or '4GiB', etc.")
-	f.BoolVar(&c.quiet, "q", false, "do not print progress updates to stdout")
 	f.BoolVar(&c.empty, "empty", false, "create an archive without initial data")
 	c.ReaderFlags.SetFlags(f)
 	return c, nil
-}
-
-func tsDir(ts nano.Ts) string {
-	return ts.Time().Format("20060102")
 }
 
 func (c *Command) Run(args []string) error {
@@ -100,48 +88,7 @@ func (c *Command) Run(args []string) error {
 	if err != nil {
 		return err
 	}
-	var w *bufwriter.Writer
-	var zw zbuf.Writer
-	var n int
-	for {
-		rec, err := reader.Read()
-		if err != nil || rec == nil {
-			if w != nil {
-				if err := w.Close(); err != nil {
-					return err
-				}
-			}
-			return err
-		}
-		if w == nil {
-			ts := rec.Ts
-			dir := filepath.Join(ark.Root, tsDir(ts))
-			if err := os.MkdirAll(dir, 0755); err != nil {
-				return err
-			}
-			path := filepath.Join(dir, ts.StringFloat()+".zng")
-			//XXX for now just truncate any existing file.
-			// a future PR will do a split/merge.
-			out, err := fs.OpenFile(path, os.O_TRUNC|os.O_WRONLY|os.O_CREATE, 0644)
-			if err != nil {
-				return err
-			}
-			if !c.quiet {
-				fmt.Printf("writing %s\n", path)
-			}
-			w = bufwriter.New(out)
-			zw = zngio.NewWriter(w, zio.WriterFlags{})
-		}
-		if err := zw.Write(rec); err != nil {
-			return err
-		}
-		n += len(rec.Raw)
-		if int64(n) >= ark.Config.LogSizeThreshold {
-			if err := w.Close(); err != nil {
-				return err
-			}
-			w = nil
-			n = 0
-		}
-	}
+	defer reader.Close()
+
+	return archive.Import(ark, reader)
 }
